Test updates

+++ tests/integrated/centered_wave/run.sh
@@ -7,11 +7,24 @@
 export I_MPI_REMOVED_VAR_WARNING=0
 export I_MPI_VAR_CHECK_SPELLING=0
 export FOR_COARRAY_NUM_IMAGES=1
+export OMP_NUM_THREADS=6
 
-rm cato.x
+cato_dir=../../../build
+run_dir=`pwd`
+python generate_ic.py
+
+if [ -f "cato.x" ]; then rm cato.x; fi
+if [ -f "cato.error" ]; then rm cato.error; fi
 rm -rf step*
-cd ../../../build &&\
-make -j &&\
-cd - &&\
-cp ../../../build/bin/cato.x . &&\
-./cato.x input.ini
+
+cd ${cato_dir} && make -j && \
+    cd ${run_dir} && \
+    cp ${cato_dir}/bin/cato.x . &&\
+    ./cato.x input.ini
+
+if [ -f "cato.error" ]; then
+    echo
+    echo "Error(s) if any"
+    echo
+    cat cato.error
+fi
